refactor(server): apis > services > daos relations
